bugfix: fixing compatibility of get_notebooks for Python 3.9

+++ src/ndp/get_notebooks.py
@@ -10,5 +10,6 @@ import pathlib
 if __name__=="__main__":
     current_directory = pathlib.Path()
 
-    for filepath in files('ndp.jupyter').joinpath('').glob('*ipynb'):
-        shutil.copy(filepath,current_directory)
+    for filepath in files('ndp.jupyter').iterdir():
+        if filepath.is_file() and ('.ipynb' in filepath.parts[-1]):
+            shutil.copy(filepath,current_directory)
